[feat]: Changed Profile and Edit Profile pages
