+++ src/components/edit-profile/edit-profile-styles.css
@@ -1,6 +1,15 @@
 .img-account-profile {
     height: 10rem;
 }
+
+.disp-inline{
+    display: inline;
+}
+
+.red-button{
+    background-color: red;
+}
+
 .rounded-circle {
     border-radius: 50% !important;
 }
