+++ src/components/profile/index.js
@@ -1,9 +1,13 @@
 import React, {useState} from "react";
 
 import ReviewList from "../review-list";
+import ProfileCard from "./profile-card";
+
 import {updateProfile} from "../../actions/userActions";
 import {useSelector} from "react-redux";
 import {Link, useNavigate} from "react-router-dom";
+
+import '../edit-profile/edit-profile-styles.css';
 
 const ProfileComponent = () => {
 
@@ -17,42 +21,41 @@ const ProfileComponent = () => {
             {
                 user ?
                 (<div className="card">
-                <img className="card-img-top" src="/images/background.jpg" alt="popcorn banner"/>
-                <div className="card-body">
-                    <div className="d-flex align-items-center">
-                        <img className="img-thumbnail rounded-circle mr-3" src="/images/user-placeholder.png"
-                             alt="person silhouette" width="150"/>
-                        <h2 className="card-title mr-3">
-                            {user.first_name} {user.last_name}
-                        </h2>
-                        <h3>
-                            <small className="text-muted">@{user.username}</small>
-                        </h3>
-                    </div>
+                <div className="container-xl px-4 mt-4">
+                    <ProfileCard/>
 
-                    <div className="ml-2">
-                        <h5 className="mt-4 mb-4">Personal information</h5>
-                        <p><i className="fa-solid fa-envelope"/> {user.email_id}</p>
-                        <p><i className="fa-solid fa-cake-candles"/> {user.DOB}</p>
-                        <p><i className="fa-solid fa-phone"/> {user.phone_no}</p>
-                    </div>
 
-                    {user.role === "ADMIN" ? <button className="btn btn-primary rounded-pill mr-3">
-                        Manage Site
-                    </button> : ''}
 
-                    <Link to='/profile/edit'
-                    //     onClick={() => {
-                    //     navigate('/profile/edit')
-                    // }}
-                        className="btn btn-secondary rounded-pill">
-                        Edit Profile
-                    </Link>
+                    {/* Extra actions for CRITIC and ADMIN */}
+                    {/*<Link to='/profile/edit'*/}
+                    {/*      className="btn btn-secondary rounded-pill">*/}
+                    {/*    Edit Profile*/}
+                    {/*</Link>*/}
 
-                    {user.role === "CRITIC" ? <ReviewList/> : ''}
+                    {user.role === "ADMIN" ?
+                        <>
+                        <br/><br/>
+                        <button className="btn btn-primary rounded">
+                        Manage Site</button>
+                        </>
+                        : ''}
+
+
+                    {user.role === "CRITIC" ?
+                        <>
+                        <div className="card mb-4 mb-xl-0">
+                            <div className="card-header"><h4>Reviews</h4>
+                            </div>
+                            <div className="card-body"><ReviewList/></div>
+                        </div>
+                        </>
+                        : ''}
+
 
                 </div>
-            </div>) : navigate("/login")
+            </div>
+
+                ) : navigate("/login")
             }
